enhance seo, basic react-helmet on all pages
