+++ package.json
@@ -29,6 +29,7 @@
     "@types/node": "^18.11.17",
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
+    "@types/react-helmet": "^6.1.11",
     "autoprefixer": "^10.4.19",
     "os-browserify": "^0.3.0",
     "postcss": "^8.4.31",
