--- src/features/about/routes/about/index.ts
@@ -1,1 +1,0 @@
-export * from "./About"
