+++ src/features/about/routes/index.ts
@@ -1,1 +1,1 @@
-export * from "./about"
+export * from "./About"
