--- src/features/contact/routes/contact/index.ts
@@ -1,1 +1,0 @@
-export * from "./Contact"
