+++ src/features/contact/routes/index.ts
@@ -1,1 +1,1 @@
-export * from "./contact";
+export * from "./Contact";
