+++ src/features/projects/routes/projectList/ProjectList.tsx
@@ -1,12 +1,22 @@
 import * as React from "react";
-import { useRecoilValue } from "recoil";
 import { Footer, MainNavbar, ScrollToTop } from "~src/components";
 import { ProjectListSection } from "../../components";
-import { projectListState } from "~src/stores";
+import { Helmet } from "react-helmet";
 
 export const ProjectList = () => {
   return (
     <>
+      <Helmet>
+        <title>Landon McKell - Projects</title>
+        <meta
+          name="description"
+          content="Explore a curated selection of Landon McKell's web development projects. From front-end interfaces to complex back-end systems, see how Landon turns ideas into functional, user-friendly applications."
+        />
+        <meta
+          name="keywords"
+          content="Landon McKell projects, web development portfolio, web design projects, software projects, front-end projects, back-end projects, full-stack projects, JavaScript projects, React projects, Next.js projects."
+        />
+      </Helmet>
       <ScrollToTop />
       <MainNavbar projects />
       <section
